Rename assemblies, removing Microsoft prefix (#5)
After discussing, we decided to remove the Microsoft prefix,
at least for now, from the assembly names. That will allow
publishing NuGet packages to nuget.org. It's also consistent
with most .NET Foundation projects, which lack a company
prefix. When we move this repo to github.com/dotnet
(assuming that's its permanent home), we may or may not
add the Microsoft prefix back.